refactor: improve pack-extension script code

+++ pack-extension.sh
@@ -3,13 +3,14 @@
 source_folder=src
 build_folder=build
 
-metadata_file=$source_folder/metadata.json
-
 extension_uuid=add-username-toppanel@brendaw.com
 
 packed_extension_file=$build_folder/$extension_uuid.zip
 
-rm -r $build_folder
+if [ -d "$build_folder" ]; then
+	rm -r $build_folder
+fi
+
 mkdir -p $build_folder
 
 cd $source_folder
